Rename BrokerAlerts to Broker and BrokerAdmin to BrokerList and update routing; enhance filter management and UI components, fix some visual issue and bad UX.
